templates/python-Flask-basic-templates/app.py
import os
from flask import Flask, request, jsonify
import json
from dotenv import load_dotenv
from flask_cors import CORS

# Load environment variables from .env file
load_dotenv()

app = Flask(__name__)
CORS(app)  # Enable CORS for all routes and origins

# Configure FLASK_DEBUG from environment variable
app.config['DEBUG'] = os.environ.get('FLASK_DEBUG')

@app.route('/')
def hello():
    return "Prueba Tecnica"

if __name__ == '__main__':
    app.run()
    